deleted constants/graphql directory and moved changes from old files to new files
