--- constants/graphql/changeInStockItemsInCart.mutation.ts
@@ -1,20 +1,0 @@
-import gql from 'graphql-tag'
-
-const CHANGE_IN_STOCK_ITEMS_IN_CART_MUTATION = gql`
-  mutation ChangeInStockItemsInCart(
-    $userId: ID!
-    $boardId: ID!
-    $quantity: Int!
-  ) {
-    changeInStockItemsInCart(
-      userId: $userId
-      boardId: $boardId
-      quantity: $quantity
-    ) {
-      id
-      quantity
-    }
-  }
-`
-
-export default CHANGE_IN_STOCK_ITEMS_IN_CART_MUTATION
