--- constants/graphql/woods.query.ts
@@ -1,13 +1,0 @@
-import gql from 'graphql-tag'
-
-const WOODS_QUERY = gql`
-  query Woods {
-      woods {
-        id
-        name
-        picture_url
-    }
-  }
-`
-
-export default WOODS_QUERY
