+++ graphql/mutations/addNewAppUser.mutation.ts
